@@ -55,8 +55,8 @@
     "history": "^4.7.2",
     "jquery": "^3.3.1",
     "lodash": "^4.17.11",
-    "react": "16.7.0-alpha.0",
-    "react-dom": "16.7.0-alpha.0",
+    "react": "^16.8.3",
+    "react-dom": "^16.8.3",
     "react-modal": "^3.8.1",
     "react-motion": "^0.5.2",
     "react-redux": "^5.1.1",
