@@ -1,21 +1,35 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
+import {Button, Modal} from 'antd'
 
-export default class HooksDemo1 extends Component {
-    state = {
-        list:[
-            {text:"在组件内部调用这个函数渲染组件",id:1},
-            {text:"在组件内部调用这个函数渲染组件",id:2},
-            {text:"在组件内部调用这个函数渲染组件",id:3},
-        ]
-    }
+function Example() {
+  // 声明一个新的状态变量，我们将其称为 "count"
+  const [count, setCount] = useState(0);
 
-    render(){
-        const { render} = this.props;
-        const {list} = this.state;
-        return (
-            <div>
-                {render(list)}
-            </div>
-        )
-    }
+  return (
+    <div>
+      <p>you clicked {count} times</p>
+      <button onClick={() => setCount(count + 1)}>click me</button>
+    </div>
+  );
 }
+
+function App() {
+    const [open, setOpen] = useState(false);
+    return (
+      <React.Fragment>
+        <Button type="primary" onClick={() => {
+            console.log("fdfd")
+            setOpen(true)
+        }}>
+          Open Modal
+        </Button>
+        <Modal
+          visible={open}
+          onOk={() => setOpen(false)}
+          onCancel={() => setOpen(false)}
+        />
+      </React.Fragment>
+    );
+  }
+
+export default Example;
